+++ CHANGELOG.md
@@ -2,6 +2,7 @@
 
 - Added `logging` category
 - Fixed missing unescaping for function name and location in the V8 log, which could lead to incorrect parsing — such as including the script path in the function name
+- Fixed `esm.sh` registry patterns
 
 ## 0.7.0 (2025-05-08)